trying CI workarounds
